+++ src/padone/common/controller/Article/ArticleListServlet.java
@@ -1,22 +1,16 @@
-package padone.common.controller;
+package padone.common.controller.Article;
 
 import javax.servlet.ServletException;
 import java.io.IOException;
-import java.io.PrintWriter;
-import java.sql.Connection;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Statement;
 
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import padone.common.model.ArticleServer;
+import padone.common.model.Article.ArticleServer;
 import com.google.gson.Gson;
 import org.apache.tomcat.jdbc.pool.DataSource;
-import org.json.JSONArray;
 
 @WebServlet("/ArticleListServlet")
 public class ArticleListServlet extends HttpServlet {
